+++ packages/block-editor/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Enhancement
+
+-  `WordBreakControl`: Add support for word break styles in the block editor.
 ## 15.3.0 (2025-09-03)
 
 ## 15.2.0 (2025-08-20)